% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qdis.R
\name{plot_qdis}
\alias{plot_qdis}
\title{Plot of the distribution of quantiles}
\usage{
plot_qdis(object_1, object_2 = NULL, ...)
}
\arguments{
\item{object_1}{Object which contains model fit}

\item{object_2}{Object which contains model fit}

\item{...}{Other arguments.}
}
\description{
Function produces plot of the distribution of quantiles for one or more model fits
}
\details{
If \code{object_2} is specified, a plot for the distribution of quantiles for \code{object}
and one for \code{object_2} will be combined in the same plot.
}
\seealso{
\code{\link{plot_qdis.lm}} and \code{\link{plot_qdis.lmvar}}
}
