% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq.R
\name{plot_qq}
\alias{plot_qq}
\title{QQ-plot}
\usage{
plot_qq(object_1, object_2 = NULL, ...)
}
\arguments{
\item{object_1}{Object which contains model fit}

\item{object_2}{Object which contains model fit}

\item{...}{Other arguments.}
}
\description{
Function produces QQ-plots for one or more model fits
}
\details{
If \code{object_2} is specified, a QQ-plot for \code{object_1} and one for \code{object_2} will be
combined in the same plot.
}
\seealso{
\code{\link{plot_qq.lm}} and \code{\link{plot_qq.lmvar}}
}
