% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq_lmlike.R
\name{plot_qq_lmlike}
\alias{plot_qq_lmlike}
\title{QQ-plot for an object of class 'lm' or class 'lmvar'}
\usage{
plot_qq_lmlike(object_1, object_2 = NULL, name_1, name_2 = NULL)
}
\arguments{
\item{object_1}{Object of class 'lm' or class 'lmvar'}

\item{object_2}{Object of class 'lm' or class 'lmvar'}

\item{name_1}{Character string, the name of \code{object_1}}

\item{name_2}{Character string, the name of \code{object_2}}
}
\description{
Function produces QQ-plots for an object of class 'lm' or 'lmvar'. This function is
called by \code{\link{plot_qq.lm}} and
\code{\link{plot_qq.lmvar}}. It is not intended to be called directly.
}
\details{
If \code{object_2} is specified, a QQ-plot for \code{object_1} and one for \code{object_2} will be
combined in the same plot.

The string \code{name_1} and (optionally) \code{name_2} are used in the legend of the plot as names for
\code{object_1} and (optionally) \code{object_2}.

All inputs of class 'lm' must contain the response vector \eqn{y}. I.e., one must run \code{\link[stats]{lm}}
with argument \code{y = TRUE}.
}
