% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cvlmvar.R
\name{print.cvlmvar}
\alias{print.cvlmvar}
\title{Print method for an object of class 'cvlmvar'}
\usage{
\method{print}{cvlmvar}(x, digits = NULL, ...)
}
\arguments{
\item{x}{Object of class 'cvlmvar'}

\item{digits}{Integer, number of significant digits of standard deviations in printed output. Must
be larger than zero, or \code{NULL}.}

\item{...}{For compatibility with \code{\link[base]{print}} generic.}
}
\description{
Print method for an object of class 'cvlmvar'. This object is created by the functions
\code{\link{cv.lm}} and \code{\link{cv.lmvar}}.
}
\details{
If \code{digits = NULL}, printed values are not rounded. Otherwise, all standard deviations are rounded to
\code{digits} significant digits. All corresponding mean values are rounded to a precision equal to the maximum
precision of the rounded value of the standard deviation.

Examples of \code{print.cvlmvar} are provided in the examples of the functions \code{\link{cv.lm}} and
\code{\link{cv.lmvar}}.
}
