% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary_lmvar.R
\name{print.summary_lmvar}
\alias{print.summary_lmvar}
\title{Print method for the summary of an 'lmvar' object.}
\usage{
\method{print}{summary_lmvar}(x, ...)
}
\arguments{
\item{x}{Object of class 'summary_lmvar'}

\item{...}{For compatibility with \code{\link[base]{print}} generic.}
}
\description{
Print method for an object of the class 'summary_lmvar'. This object is created by \code{\link{summary.lmvar}}.
}
\seealso{
\code{\link{summary.lmvar}} for a summary of the fit present in an object of class 'lmvar'.
}
